+++ gitcmd/gitcmd.py
@@ -49,6 +49,25 @@ def in_repository(path, ignore=True):
     return ret
 
 
+def top_level(path):
+    """Return the absolute path of the top-level directory."""
+    if not in_repository(path):
+        raise NotInRepositoryError("'" + path + "' is not inside a repository")
+    
+    cwd = os.getcwd()
+    
+    try:
+        os.chdir(path) if os.path.isdir(path) else os.chdir(os.path.dirname(path))
+        cmd = "git rev-parse --show-toplevel"
+        p = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+        out, err = p.communicate()
+    finally:
+        os.chdir(cwd)
+    
+    return (p.returncode, out.decode()[:-1], err.decode())
+    
+
+
 def remote_url(path, remote='origin'):
     """Return the remote's URL (default 'origin') of the repository pointed by path."""
     if not in_repository(path):